ShareSplitter: Introduce class
Add a new `ShareSplitter` class that uses the Yahoo Finance API (through
the `yfinance` package) to get the list of all share sub-division and
share consolidation events that took place for the tickers referred in
the orders parsed. Once retrieved, this information is persisted on a
local cache to avoid unnecessary remote procedure calls and the Yahoo
Finance API is only used again if needed.

In a following commit, this class will also be able to adjust the
quantity and amount attributed to an order based on the share splits
that apply to that order.