Use ISIN as the primary means to identify a security
Update the codebase to use the ISIN as the primary means to identify a
security as opposite to using the ticker symbol. The later is easier to
memorise, but the same security might trade using different tickers on
different exchanges. An example is Unilever Plc (GB00B10RZP78) which
trades in the Euronext Amsterdam as "UNA" and as "ULVR" in the London
Stock Exchange. If shares were bought on both exchanges, the capital
gains would be incorrect if calculated based on the ticker symbol.