findata: Use non-magic method to retrieve security information
Instead of providing a container interface to retrieve security
information, add a non-magic method for that purpose. In a following
commit, the `FinancialData` class will be extended to provide currency
exchange rates as well, and it is less appropriate to provide a
container interface that serves two different types of information.