Tests: Split lines before comparing output
Split lines before comparing captured output against expected output. On
Microsoft Windows, the CSV output will use "\r\n" as line separator,
which is different than the "\n" stored in the files containing the
expected output. By using `.splitlines()` the platform-dependent line
separators are removed and the content can be compared for equality
independently of in which platform was produced.