PrettyTable: Add new Field dataclass to describe a field
Add a `Field` dataclass with all metadata needed for a field instead of
forcing clients to repeat the field name more than once in order to set
field attributes. The alignment for a field is now controlled using an
attribute as well instead of inferring it from the field name.