Replace iso4217 package with py-moneyed
The moneyed package provides a `Money` type that will be needed in a
future commit to support multi-currency accounts. As this package also
provides a `Currency` type, the iso4217 package is now providing
duplicated functionality and this commit migrates away from the latter.
With a `Money` type, the `Price` type is also not needed and it is also
removed.