PrettyTable: Add support for a Money objects
Add support for `Money` objects. `Money` objects need special handling
as they can refer to different currencies. This commit implements the
following behaviour for text and HTML exports:

1. If there are money values in different currencies, both the amount
   and the currency code are displayed in each row.
2. If all money values are in the same currency, only the amount is
   displayed in each row (to reduce visual noise) and the currency code
   is displayed in the field title instead.

CSV and JSON exports always use a separate field to indicate the
currency code. To this end, the `get_csv()` and `get_json()` methods
need to be overriden.