Add support for multi-currency accounts
Add support for multi-currency accounts where transaction totals might
be associated with different currencies. This commit enables to view
these transactions in the orders, dividends, transfers and interest
subcommands. However, at the moment the `TaxCalculator` class does not
support calculating capital gains in orders whose total is not in GBP.
Therefore, we bail out if calculation of capital gains is needed and
these type of orders are found.

Fixes #40.