Change how the total field is set for an order
At the moment, the `total` field for an order is set in a way so that:

1. It is *not* inclusive of fees for an acquisition.
2. Contains the gross proceeds (before fees are deducted) for a
   disposal.

After this commit the `total` field:

1. Becomes inclusive of fees for an acquisition.
2. Contains the net proceeds (after fees being deducted) for a disposal.

This aligns in how Trading212 and Freetrade set this field in the CSV.
The `total_cost` and `net_proceeds` properties are now redundant and are
removed in favour of using the `total` field directly. The support to
disregard forex fees as an allowable cost is temporarily removed but it
will be added back in a following commit.